refactor(userspace/libsinsp)!: drop gen_event filter interfaces
