+++ userspace/libsinsp/eventformatter.h
@@ -23,7 +23,7 @@ limitations under the License.
 #include <json/json.h>
 
 #include <libsinsp/filter_check_list.h>
-#include <libsinsp/gen_filter.h>
+#include <libsinsp/filter.h>
 
 class sinsp_filter_check;
 
@@ -36,9 +36,14 @@ class sinsp_filter_check;
   This class can be used to format an event into a string, based on an arbitrary
   format.
 */
-class SINSP_PUBLIC sinsp_evt_formatter : public gen_event_formatter
+class SINSP_PUBLIC sinsp_evt_formatter
 {
 public:
+	enum output_format {
+		OF_NORMAL = 0,
+		OF_JSON   = 1
+	};
+
 	/*!
 	  \brief Constructs a formatter.
 
@@ -51,8 +56,6 @@ public:
 	sinsp_evt_formatter(sinsp* inspector, filter_check_list &available_checks);
 
 	sinsp_evt_formatter(sinsp* inspector, const std::string& fmt, filter_check_list &available_checks);
-
-	void set_format(gen_event_formatter::output_format of, const std::string& fmt) override;
 
 	virtual ~sinsp_evt_formatter();
 
@@ -68,13 +71,15 @@ public:
 	*/
 	bool resolve_tokens(sinsp_evt *evt, std::map<std::string,std::string>& values);
 
-	// For compatibility with gen_event_filter_factory
+	virtual void set_format(output_format of, const std::string& fmt);
+
+	// For compatibility with sinsp_filter_factory
 	// interface. It just calls resolve_tokens().
-	bool get_field_values(gen_event *evt, std::map<std::string, std::string> &fields) override;
+	virtual bool get_field_values(sinsp_evt *evt, std::map<std::string, std::string> &fields);
 
-	void get_field_names(std::vector<std::string> &fields) override;
+	virtual void get_field_names(std::vector<std::string> &fields);
 
-	gen_event_formatter::output_format get_output_format() override;
+	virtual output_format get_output_format();
 
 	/*!
 	  \brief Fills res with the string rendering of the event.
@@ -85,12 +90,17 @@ public:
 	  \return true if the string should be shown (based on the initial *),
 	   false otherwise.
 	*/
-	bool tostring(sinsp_evt* evt, OUT std::string* res);
+	inline bool tostring(sinsp_evt* evt, OUT std::string* res)
+	{
+		if (!res)
+		{
+			return false;
+		}
+		return tostring(evt, *res);
+	}
+	virtual bool tostring(sinsp_evt* evt, std::string &output);
 
-	// For compatibility with gen_event_formatter
-	bool tostring(gen_event* evt, std::string &output) override;
-
-	bool tostring_withformat(gen_event* evt, std::string &output, gen_event_formatter::output_format of) override;
+	virtual bool tostring_withformat(sinsp_evt* evt, std::string &output, output_format of);
 
 	/*!
 	  \brief Fills res with end of capture string rendering of the event.
@@ -102,7 +112,7 @@ public:
 	bool on_capture_end(OUT std::string* res);
 
 private:
-	gen_event_formatter::output_format m_output_format;
+	output_format m_output_format;
 
 	// vector of (full string of the token, filtercheck) pairs
 	// e.g. ("proc.aname[2], ptr to sinsp_filter_check_thread)
@@ -117,22 +127,22 @@ private:
 	Json::FastWriter m_writer;
 };
 
-class sinsp_evt_formatter_factory : public gen_event_formatter_factory
+class sinsp_evt_formatter_factory
 {
 public:
 	sinsp_evt_formatter_factory(sinsp *inspector, filter_check_list &available_checks);
 	virtual ~sinsp_evt_formatter_factory() = default;
 
-	void set_output_format(gen_event_formatter::output_format of) override;
+	virtual void set_output_format(sinsp_evt_formatter::output_format of);
 
-	std::shared_ptr<gen_event_formatter> create_formatter(const std::string &format) override;
+	virtual std::shared_ptr<sinsp_evt_formatter> create_formatter(const std::string &format);
 
 protected:
 
 	// Maps from output string to formatter
-	std::map<std::string, std::shared_ptr<gen_event_formatter>> m_formatters;
+	std::map<std::string, std::shared_ptr<sinsp_evt_formatter>> m_formatters;
 
 	sinsp *m_inspector;
 	filter_check_list &m_available_checks;
-	gen_event_formatter::output_format m_output_format;
+	sinsp_evt_formatter::output_format m_output_format;
 };
