+++ userspace/libsinsp/sinsp_filtercheck_event.h
@@ -97,8 +97,6 @@ public:
 	Json::Value extract_as_js(sinsp_evt*, OUT uint32_t* len) override;
 	bool compare(sinsp_evt*) override;
 
-	void validate_filter_value(const char* str, uint32_t len);
-
 	uint64_t m_u64val;
 	int64_t m_s64val;
 	uint64_t m_tsdelta;
@@ -117,6 +115,7 @@ public:
 	filtercheck_field_info m_customfield;
 
 private:
+	void validate_filter_value(const char* str, uint32_t len);
 	int32_t extract_arg(std::string fldname, std::string val, OUT const struct ppm_param_info** parinfo);
 	int32_t extract_type(std::string fldname, std::string val, OUT const struct ppm_param_info** parinfo);
 	uint8_t* extract_error_count(sinsp_evt *evt, OUT uint32_t* len);
