+++ userspace/libsinsp/test/filter_compiler.ut.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 // A mock filtercheck that returns always true or false depending on
 // the passed-in field name. The operation is ignored.
-class mock_compiler_filter_check : public gen_event_filter_check
+class mock_compiler_filter_check : public sinsp_filter_check
 {
 public:
 	inline int32_t parse_field_name(const char* str, bool a, bool n) override
@@ -16,7 +16,7 @@ public:
 		return 0;
 	}
 
-	inline bool compare(gen_event*) override
+	inline bool compare(sinsp_evt*) override
 	{
 		if (m_name == "c.true")
 		{
@@ -42,7 +42,7 @@ public:
 		m_value = string(str, l);
 	}
 
-	inline bool extract(gen_event *e, OUT vector<extract_value_t>& v, bool) override
+	inline bool extract(sinsp_evt *e, OUT vector<extract_value_t>& v, bool) override
 	{
 		return false;
 	}
@@ -52,28 +52,28 @@ public:
 };
 
 // A factory that creates mock filterchecks
-class mock_compiler_filter_factory: public gen_event_filter_factory
+class mock_compiler_filter_factory: public sinsp_filter_factory
 {
 public:
-	mock_compiler_filter_factory(sinsp *inspector) : m_inspector(inspector) {}
+	mock_compiler_filter_factory(sinsp *inspector): sinsp_filter_factory(inspector, m_filterlist) {}
 
-	inline gen_event_filter *new_filter() override
+	inline sinsp_filter *new_filter() override
 	{
 		return new sinsp_filter(m_inspector);
 	}
 
-	inline gen_event_filter_check *new_filtercheck(const char *fldname) override
+	inline sinsp_filter_check *new_filtercheck(const char *fldname) override
 	{
 		return new mock_compiler_filter_check();
 	}
 
-	inline list<gen_event_filter_factory::filter_fieldclass_info> get_fields() const override
+	inline list<sinsp_filter_factory::filter_fieldclass_info> get_fields() const override
 	{
 		return m_list;
 	}
 
-	list<gen_event_filter_factory::filter_fieldclass_info> m_list;
-	sinsp *m_inspector;
+	sinsp_filter_check_list m_filterlist;
+	list<sinsp_filter_factory::filter_fieldclass_info> m_list;
 };
 
 // Compile a filter, pass a mock event to it, and
@@ -82,7 +82,7 @@ public:
 void test_filter_run(bool result, string filter_str)
 {
 	sinsp inspector;
-	std::shared_ptr<gen_event_filter_factory> factory;
+	std::shared_ptr<sinsp_filter_factory> factory;
 	factory.reset(new mock_compiler_filter_factory(&inspector));
 	sinsp_filter_compiler compiler(factory, filter_str);
 	try
@@ -101,7 +101,7 @@ void test_filter_run(bool result, string filter_str)
 }
 
 void test_filter_compile(
-		std::shared_ptr<gen_event_filter_factory> factory,
+		std::shared_ptr<sinsp_filter_factory> factory,
 		string filter_str,
 		bool expect_fail=false)
 {
@@ -181,7 +181,7 @@ TEST(sinsp_filter_compiler, str_escape)
 TEST(sinsp_filter_compiler, supported_operators)
 {
 	sinsp inspector;
-	std::shared_ptr<gen_event_filter_factory> factory(new mock_compiler_filter_factory(&inspector));
+	std::shared_ptr<sinsp_filter_factory> factory(new mock_compiler_filter_factory(&inspector));
 
 	// valid operators
 	test_filter_compile(factory, "c.true exists");
@@ -216,7 +216,7 @@ TEST(sinsp_filter_compiler, supported_operators)
 TEST(sinsp_filter_compiler, complex_filter)
 {
 	sinsp inspector;
-	std::shared_ptr<gen_event_filter_factory> factory(new mock_compiler_filter_factory(&inspector));
+	std::shared_ptr<sinsp_filter_factory> factory(new mock_compiler_filter_factory(&inspector));
 
 	// This is derived from the Falco default rule
 	// "Unexpected outbound connection destination" coming from here:
